refactor: Convert affiliate earnings payout model to typescript
